+++ routes/index.tsx
@@ -1,11 +1,12 @@
 import { DateTime, datetime, diffInDays } from 'ptera';
 import { WithSession } from 'fresh_session';
 import { Handlers, PageProps } from '$fresh/server.ts';
-import { Card, Footer, Layout, Link, Main, Navigation, Text } from 'lunchbox';
+import { Card, Footer, Layout, Link, Main, Text } from 'lunchbox';
 import { dbEntry, getAllEntries } from 'db/entry.ts';
 import { dbTag, getAllTags, iTag } from 'db/tag.ts';
 import { EntryInput } from 'islands/EntryInput/index.tsx';
 import { EntryList } from 'components/EntryList/index.tsx';
+import Navigation from 'components/Navigation/index.tsx';
 import { redirect } from 'redirect';
 import { getApp, iApp } from 'db/index.ts';
 import { createDictionaryDocument } from 'utils';
@@ -65,14 +66,7 @@ export default function Home(props: PageProps<HomePageData>) {
 
   return (
     <>
-      <Navigation class='py-3'>
-        <Layout dashboard type='left'>
-          <Text>{today.format('MMM d')}</Text>
-          <div class='flex gap-6'>
-            <Link href='/archive'>Archive</Link>
-          </div>
-        </Layout>
-      </Navigation>
+      <Navigation today={today} currentPage='home' />
       <Main
         data-starting_utc_date={appConfig?.startingUtcDate}
         class='min-h-screen mt-10 flex flex-col gap-9'